Cloudrun_Function/requirements.txt
functions-framework==3.*
google-cloud-storage
pandas
cloud-sql-python-connector
pymysql
sqlalchemy
numpy
python-dotenv
requests